// backend/src/csv/dto/csv.dto.ts
// Define the CSVDto interface
export class CSVDto {
  name: string;
  ticker: string;
  nomicsid: string;
  coingeckoid: string;
  statuses: string;
  statusreasoning: string;
  primarysector: string;
  secondarysector: string;
  facts_data: string;
  opinions: string;
  rm_tr: string;
  format: string;
  coingeckolink: string;
  assetdescription: string;
  tokendescription: string;
  watchlist: string;
  messarilink: string;
}